pyproject.toml
[tool.poetry]
name = "powerpoint-automation"
version = "0.1.3"
description = ""
authors = ["Patrick Stoeckle <patrick.stoeckle@tum.de>"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
click = "^7.1.2"
python-pptx = "^0.6.18"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-broken-line = "^0.3.0"
flake8-bugbear = "^20.11.1"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.3.1"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^1.0.0"
flake8-gl-codeclimate = "^0.1.4"
flake8-html = "^0.4.1"
flake8-isort = "^4.0.0"
flake8-pytest-style = "^1.3.0"
flake8-rst-docstrings = "^0.0.14"
flake8-string-format = "^0.3.0"
mypy = "^0.812"
lxml = "^4.6.3"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
powerpoint-automation = 'powerpoint_automation.main:main_group'
